requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-health-check==3.17.0
psycopg2==2.9.7
PyJWT==2.8.0
requests==2.31.0
drf-writable-nested==0.7.0
cryptography==41.0.3
django-redis==5.3.0
gunicorn==21.2.0
boto3==1.28.50